Move DirtyDiffModel to avoid cyclic import
