+++ src/vs/workbench/contrib/scm/browser/scm.contribution.ts
@@ -40,6 +40,7 @@ import { isSCMRepository } from './util.js';
 import { SCMHistoryViewPane } from './scmHistoryViewPane.js';
 import { IsWebContext } from '../../../../platform/contextkey/common/contextkeys.js';
 import { RemoteNameContext } from '../../../common/contextkeys.js';
+import { DirtyDiffModelService, IDirtyDiffModelService } from './diff.js';
 
 ModesRegistry.registerLanguage({
 	id: 'scminput',
@@ -596,3 +597,4 @@ KeybindingsRegistry.registerCommandAndKeybindingRule({
 registerSingleton(ISCMService, SCMService, InstantiationType.Delayed);
 registerSingleton(ISCMViewService, SCMViewService, InstantiationType.Delayed);
 registerSingleton(IQuickDiffService, QuickDiffService, InstantiationType.Delayed);
+registerSingleton(IDirtyDiffModelService, DirtyDiffModelService, InstantiationType.Delayed);
